+++ kowalski/src/main.rs
@@ -3,6 +3,7 @@ use clap::Parser;
 mod horde;
 mod http_api;
 mod http_ops;
+mod rookery;
 
 #[derive(Parser, Debug)]
 #[clap(